Move anonymize package
