+++ client/anonymize/anonymize.go
