+++ client/anonymize/anonymize_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/netbirdio/netbird/util/anonymize"
+	"github.com/netbirdio/netbird/client/anonymize"
 )
 
 func TestAnonymizeIP(t *testing.T) {
