+++ client/cmd/status.go
@@ -16,11 +16,11 @@ import (
 	"google.golang.org/grpc/status"
 	"gopkg.in/yaml.v3"
 
+	"github.com/netbirdio/netbird/client/anonymize"
 	"github.com/netbirdio/netbird/client/internal"
 	"github.com/netbirdio/netbird/client/internal/peer"
 	"github.com/netbirdio/netbird/client/proto"
 	"github.com/netbirdio/netbird/util"
-	"github.com/netbirdio/netbird/util/anonymize"
 	"github.com/netbirdio/netbird/version"
 )
